Add the store method to a PostRequests trait to share functionality between routes
