+++ app/Http/Controllers/PostController.php
@@ -2,14 +2,15 @@
 
 namespace Rogue\Http\Controllers;
 
-use Illuminate\Http\Request;
 use Rogue\Services\PostService;
-use Rogue\Http\Requests\PostRequest;
 use Rogue\Repositories\SignupRepository;
 use Rogue\Http\Transformers\PostTransformer;
+use Rogue\Http\Controllers\Traits\PostRequests;
 
 class PostController extends Controller
 {
+    use PostRequests;
+
     /**
      * The post service instance.
      *
@@ -44,45 +45,6 @@ class PostController extends Controller
         $this->signups = $signups;
 
         $this->transformer = new PostTransformer;
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    // @TODO - Move to shared Post Trait so that
-    public function store(PostRequest $request)
-    {
-        $transactionId = incrementTransactionId($request);
-
-        $signup = $this->signups->get($request['northstar_id'], $request['campaign_id'], $request['campaign_run_id']);
-
-        $updating = ! is_null($signup);
-
-        // @TODO - should we eventually throw an error if a signup doesn't exist before a post is created? I create one here because we haven't implemented sending signups to rogue yet, so it will have to create a signup record for all posts.
-        if (! $updating) {
-            $signup = $this->signups->create($request->all());
-
-            // Send the data to the PostService class which will handle determining
-            // which type of post we are dealing with and which repostitory to use to actually create the post.
-            $post = $this->posts->create($request->all(), $signup->id, $transactionId);
-
-            $code = 200;
-
-            return $this->item($post);
-        } else {
-            $post = $this->posts->update($signup, $request->all(), $transactionId);
-
-            $code = 201;
-
-            if (isset($request['file'])) {
-                return $this->item($post);
-            } else {
-                return $signup;
-            }
-        }
     }
 
     /**
